requirements.txt
Flask==3.0.3
marshmallow==4.1.0
Werkzeug==3.1.3
cryptography
python-dotenv==1.2.1
flask-cors==6.0.1
anytree==2.13.0
lark-parser==0.12.0
gunicorn==20.1.0
automata-lib==9.1.2
firebase_admin==7.1.0